Changed farm data to sqlite + water bugfix
Sqlite change is that the data used to store tillable farmland and
farmed farmland, as well as planted seeds in their respective GUI
locations, was a little bit awkward to work with. This has now been
changed to be on sqlite instead. Possible to design something in the
future that shows this working in a GUI way. Water change is that
soil would not be autowatered when the game started with rain, this
has now been fixed. Added Python version to .gitignore. It's 3.12.